modified: nmrih_map2fileserver.sh, modified: nmrih_map2game, comment_by_gitpush.sh

+++ nmrih_map2fileserver.sh
@@ -8,14 +8,18 @@ echo fname=$fname
 cd ~nmrihserver
 ls -laF $1
 unzip $1
-mv $fname.bsp $fname.nav $fname.nmo /var/www/html/maps/
+if [ -e ${fname}.bsp ];then
+	mv $fname.bsp $fname.nav $fname.nmo /var/www/html/maps/
+else
+	mv $fname/$fname.bsp $fname/$fname.nav $fname/$fname.nmo /var/www/html/maps/
+fi
 cd /var/www/html/maps/
 bzip2 $fname.*
 bzip2 -d $fname.*
 bzip2 $fname.*
 
 echo ---------------
-ls -laF /var/www/html/maps/${fname}.*
+ls -l /var/www/html/maps/${fname}.*
 echo ---------------
-ls -laF ~nmrihserver
+ls -l ~nmrihserver
 echo end.
